# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform V2 API",
...
    "version": "0.5.0"
  },
  "paths": {
...
    "/v1/scans": {
      "post": {
        "tags": [
          "scans"
        ],
        "summary": "Create scan and enqueue collection",
        "description": "Starts a new security scan for the given integration. Enqueues a job on `platform:collect.aws`; workers collect AWS inventory to S3, ingest assets, evaluate policies, and compute CIS score.\n\nRequires **tenant_admin** or **super_admin**. Poll **GET /v1/scans/{id}** until status is `completed`, `completed_with_errors`, or `failed` before reading findings.",
        "operationId": "create_scan_v1_scans_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanResponse"
                }
              }
            }
          },
          "404": {
            "description": "Integration not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "List scans",
        "description": "Recent scans for the tenant, newest first.",
        "operationId": "list_scans_v1_scans_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScanResponse"
                  },
                  "title": "Response List Scans V1 Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/scans/{scan_id}/timeline": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Scan event timeline",
        "description": "Append-only collection/policy events for the scan (from `assets.collection_events`).",
        "operationId": "get_scan_timeline_v1_scans__scan_id__timeline_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Timeline events",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "title": "Response Get Scan Timeline V1 Scans  Scan Id  Timeline Get"
                }
              }
            }
          },
          "404": {
            "description": "Scan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/scans/{scan_id}": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Get scan by ID",
        "description": "Scan status, account_id, collection sub-status, and error payload if failed.",
        "operationId": "get_scan_v1_scans__scan_id__get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scan detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanDetailResponse"
                }
              }
            }
          },
          "404": {
            "description": "Scan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/scans/{scan_id}/compliance": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "CIS compliance summary",
        "description": "Weighted CIS AWS v6 score and control summary for the scan. Same data as compliance framework endpoint, compact shape.",
        "operationId": "get_scan_compliance_summary_v1_scans__scan_id__compliance_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          },
          {
            "name": "framework_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "cis_aws_v6",
              "title": "Framework Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compliance summary",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Scan Compliance Summary V1 Scans  Scan Id  Compliance Get"
                }
              }
            }
          },
          "404": {
            "description": "Scan or compliance results not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/scans/{scan_id}/inventory-completeness": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Inventory completeness score",
        "description": "Ops metric: expected vs collected resource types for the scan.",
        "operationId": "get_inventory_completeness_v1_scans__scan_id__inventory_completeness_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Completeness report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Inventory Completeness V1 Scans  Scan Id  Inventory Completeness Get"
                }
              }
            }
          },
          "404": {
            "description": "Scan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/scans/{scan_id}/policy-coverage": {
      "get": {
        "tags": [
          "scans"
        ],
        "summary": "Policy evaluation coverage",
        "description": "Which of the ~35 YAML policies were evaluated vs skipped for this scan.",
        "operationId": "get_policy_coverage_v1_scans__scan_id__policy_coverage_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Policy coverage report",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Policy Coverage V1 Scans  Scan Id  Policy Coverage Get"
                }
              }
            }
          },
          "404": {
            "description": "Scan not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/compliance/frameworks/{framework_id}/scans/{scan_id}": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "CIS control matrix for scan",
        "description": "Full compliance result: score, per-control status, linked findings. Used by platform-ui CIS tab.",
        "operationId": "get_scan_compliance_v1_compliance_frameworks__framework_id__scans__scan_id__get",
        "parameters": [
          {
            "name": "framework_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Framework Id"
            }
          },
          {
            "name": "scan_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Scan Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Compliance matrix",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScanComplianceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Results not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/admin/scans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List scans (cross-tenant)",
        "description": "Recent scans across all tenants. Filter `?status=failed` for ops triage.",
        "operationId": "list_admin_scans_v1_admin_scans_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 200,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminScanResponse"
                  },
                  "title": "Response List Admin Scans V1 Admin Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/admin/tenants/{tenant_id}/scans": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List scans for tenant (admin)",
        "description": "Recent scans for one tenant from admin-ui tenant detail.",
        "operationId": "list_tenant_scans_v1_admin_tenants__tenant_id__scans_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminScanResponse"
                  },
                  "title": "Response List Tenant Scans V1 Admin Tenants  Tenant Id  Scans Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      },
      "post": {
        "tags": [
          "admin"
        ],
        "summary": "Trigger scan for tenant (admin)",
        "description": "Ops-initiated scan from admin-ui tenant detail. Same pipeline as POST /v1/scans.",
        "operationId": "create_tenant_scan_v1_admin_tenants__tenant_id__scans_post",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminScanCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Scan queued",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminScanCreateResponse"
                }
              }
            }
          },
          "404": {
            "description": "Tenant or integration not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "AdminScanCreate": {
        "properties": {
          "integration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Integration Id"
          }
        },
        "type": "object",
        "required": [
          "integration_id"
        ],
        "title": "AdminScanCreate"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScanComplianceResponse": {
        "properties": {
          "framework_id": {
            "type": "string",
            "title": "Framework Id"
          },
          "framework_title": {
            "type": "string",
            "title": "Framework Title"
          },
          "version_label": {
            "type": "string",
            "title": "Version Label"
          },
          "scan_id": {
            "type": "string",
            "title": "Scan Id"
          },
          "score": {
            "type": "number",
            "title": "Score",
            "description": "Weighted compliance score 0\u2013100"
          },
          "summary": {
            "additionalProperties": true,
            "type": "object",
            "title": "Summary",
            "description": "pass/fail/not_assessed counts"
          },
          "evaluated_at": {
            "type": "string",
            "title": "Evaluated At"
          },
          "controls": {
            "items": {
              "$ref": "#/components/schemas/ControlResultResponse"
            },
            "type": "array",
            "title": "Controls"
          }
        },
        "type": "object",
        "required": [
          "framework_id",
          "framework_title",
          "version_label",
          "scan_id",
          "score",
          "summary",
          "evaluated_at",
          "controls"
        ],
        "title": "ScanComplianceResponse"
      },
      "ScanCreate": {
        "properties": {
          "integration_id": {
            "type": "string",
            "format": "uuid",
            "title": "Integration Id",
            "description": "UUID from GET /v1/integrations"
          }
        },
        "type": "object",
        "required": [
          "integration_id"
        ],
        "title": "ScanCreate"
      },
      "ScanDetailResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "integration_id": {
            "type": "string",
            "title": "Integration Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Id"
          },
          "collection_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Collection Status"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "integration_id",
          "status",
          "trace_id",
          "created_at",
          "updated_at"
        ],
        "title": "ScanDetailResponse"
      },
      "ScanResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "integration_id": {
            "type": "string",
            "title": "Integration Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "integration_id",
          "status",
          "trace_id",
          "created_at",
          "updated_at"
        ],
        "title": "ScanResponse"
      },
...
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase `access_token` from POST .../auth/v1/token?grant_type=password. Paste token only in Authorize dialog."
      },
      "DevTenantId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant UUID. Only when DEV_TENANT_HEADER_AUTH=true."
      },
      "DevRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Role",
        "description": "tenant_admin | viewer | super_admin. Pair with DevTenantId."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local dev"
    },
    {
      "url": "https://api.example.com",
      "description": "Staging / production"
    }
  ]
}